# architecture/openapi-fcom-curation.yaml
openapi: 3.0.0
info:
  title: FCOM Curation & Management API
  version: 1.0.0
  description: |
    Centralized REST API for managing FCOM rule files across multiple Unified Assurance (UA) server environments.
    Interfaces with each UA server's native REST API and SVN version control system.
servers:
  - url: https://fcom-curator.example.com/api/v1
    description: Production API
...
paths:
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Authenticate and create a session
      description: |
        Supports two authentication methods against the specified UA server:
        1. Basic Auth: username + password (HTTP)
        2. Certificate Auth: TLS certificate + key pair
        Returns a session token and HTTP-only cookie for subsequent requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required:
                    - server_id
                    - auth_type
                    - username
                    - password
                  properties:
                    server_id:
                      type: string
                      example: "prod-ua-01"
                      description: Unique identifier for the UA server instance
                    auth_type:
                      type: string
                      enum: [basic]
                      description: Authentication method
                    username:
                      type: string
                      example: "api"
                      description: Username on the target UA server
                    password:
                      type: string
                      format: password
                      description: Password for the user on the target UA server
                - type: object
                  required:
                    - server_id
                    - auth_type
                    - cert_path
                    - key_path
                    - ca_cert_path
                  properties:
                    server_id:
                      type: string
                    auth_type:
                      type: string
                      enum: [certificate]
                      description: Certificate-based TLS authentication
                    cert_path:
                      type: string
                      example: "/etc/ssl/User-api.crt"
                      description: Path to user certificate file
                    key_path:
                      type: string
                      example: "/etc/ssl/User-api.key"
                      description: Path to private key file
                    ca_cert_path:
                      type: string
                      example: "/etc/ssl/BundleCA.crt"
                      description: Path to CA certificate bundle
      responses:
        '200':
          description: Session created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                    description: Opaque session token (also set as HTTP-only cookie)
                  user:
                    type: string
                  server_id:
                    type: string
                  auth_method:
                    type: string
                    enum: [basic, certificate]
                  expires_at:
                    type: string
                    format: date-time
        '401':
          description: Invalid credentials or certificate
        '400':
          description: Missing required fields or invalid server_id
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Terminate session
      security:
        - sessionCookie: []
      responses:
        '204':
          description: Session terminated
  /auth/session:
    get:
      tags:
        - Authentication
      summary: Get current session details
      security:
        - sessionCookie: []
      responses:
        '200':
          description: Current session info
          content:
            application/json:
              schema:
                type: object
                properties:
                  session_id:
                    type: string
                  user:
                    type: string
                  server_id:
                    type: string
                  server_name:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
...
  /files/browse:
    get:
      tags:
        - File Browser
      summary: List directory tree for FCOM files
      description: |
        Fetches the file/directory structure from the active UA server's SVN repository.
        Supports filtering by vendor, protocol (trap, corba, etc.), and search queries.
      security:
        - sessionCookie: []
      parameters:
        - name: path
          in: query
          schema:
            type: string
            default: "/"
          description: Path within /coms to browse (e.g., "/trap/cisco")
        - name: vendor
          in: query
          schema:
            type: string
          description: Filter by vendor folder (e.g., "cisco", "arista")
        - name: protocol_type
          in: query
          schema:
            type: string
            enum: [trap, corba, webhook, default, syslog]
          description: Filter by protocol/type subdirectory
        - name: search
          in: query
          schema:
            type: string
          description: Search term for file names or content metadata
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Directory listing
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        type:
                          type: string
                          enum: [file, directory]
                        path:
                          type: string
                        size:
                          type: integer
                          description: File size in bytes (null for directories)
                        last_modified:
                          type: string
                          format: date-time
                        last_author:
                          type: string
                          description: SVN commit author
  /files/{file_id}/preview:
    get:
      tags:
        - File Browser
      summary: Quick preview of a file (structure outline only)
      description: Returns minimal metadata and a JSON outline without full parsing.
      security:
        - sessionCookie: []
      parameters:
        - name: file_id
          in: path
          required: true
          schema:
            type: string
            description: URL-encoded file path (e.g., "trap%2Fcisco%2FCISCO-TRAP-001.json")
      responses:
        '200':
          description: File preview
          content:
            application/json:
              schema:
                type: object
                properties:
                  file_id:
                    type: string
                  path:
                    type: string
                  size:
                    type: integer
                  last_modified:
                    type: string
                    format: date-time
                  last_author:
                    type: string
                  object_count:
                    type: integer
                    description: Count of objects in the 'objects' array
                  objects_preview:
                    type: array
                    items:
                      type: object
                      properties:
                        "@objectName":
                          type: string
                        description:
                          type: string
                          maxLength: 150
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: FCOM_SESSION_ID
      description: HTTP-only session cookie set after successful login
...
security:
  - sessionCookie: []